==> 0_schedule.ipynb <==

import schedule
import time
import subprocess
import os



def run_script(script_path):
    """Runs the script using subprocess."""
    try:
        print(f"Running script: {script_path} at {time.strftime('%Y-%m-%d %H:%M:%S')}")
        process = subprocess.Popen(['python', script_path], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
        stdout, stderr = process.communicate()
        if stdout:
            print("Script output:\n", stdout.decode())
        if stderr:
            print("Script errors:\n", stderr.decode())
    except FileNotFoundError:
        print(f"Error: Script not found at {script_path}")
    except Exception as e:
        print(f"An error occurred while running {script_path}: {e}")



script1_path = "1_fetch.ipynb"  
schedule.every().day.at("10:00").do(run_script, script_path=script1_path)

script2_path = "2_analyse.ipynb"  
schedule.every().day.at("10:15").do(run_script, script_path=script2_path)

script3_path = "3_forward.ipynb" 
schedule.every().day.at("10:30").do(run_script, script_path=script3_path)

script4_path = "4_upload.ipynb"  
schedule.every().day.at("10:40").do(run_script, script_path=script4_path)


while True:
    schedule.run_pending()
    time.sleep(1)
