==> 1_fetch.ipynb <==
import pandas as pd
import re
import string
from nltk.corpus import stopwords
import nltk
import emoji
import imaplib
import email
from email.header import decode_header
import pandas as pd
from datetime import datetime, timedelta
import pytz
import configparser
import json
import quopri

config = configparser.ConfigParser()
config.read("app.config")


IMAP_SERVER = config["fetch"]["IMAP_SERVER"]
SMTP_SERVER = config["fetch"]["SMTP_SERVER"]
EMAIL_ACCOUNT = config["fetch"]["EMAIL_ACCOUNT"]
APP_PASSWORD = config["fetch"]["APP_PASSWORD"]

mail = imaplib.IMAP4_SSL(IMAP_SERVER)
mail.login(EMAIL_ACCOUNT, APP_PASSWORD)
mail.select('inbox')

since_date = (datetime.utcnow() - timedelta(hours=24)).strftime('%d-%b-%Y')
result, data = mail.uid('search', None, f'(SINCE {since_date})')

print(f'type of data is {type(data)}')
print(data)

print(f'type of result is {type(result)}')
print(result)

uids = data[0].split()

print(f'type of uids is {type(uids)}')
print(uids)


gitdef decode_mime_words(s):
    """
    A function that explicitly handles Quoted-Printable encoding for UTF-8, 
    especially for characters like emojis.
    """
    if not s:
        return ""
    decoded = email.header.decode_header(s)
    decoded_parts = []
    for part, charset in decoded:
        if isinstance(part, bytes):
            if charset and charset.lower() == 'utf-8' and b'=' in part:
                try:
                    decoded_part = quopri.decode(part).decode('utf-8')
                except UnicodeDecodeError:
                    decoded_part = part.decode('utf-8', errors='ignore')
            else:
                decoded_part = part.decode(charset or 'utf-8', errors='ignore')
            decoded_parts.append(decoded_part)
        else:
            decoded_parts.append(str(part)) 

    return ''.join(decoded_parts)



emails = []
cutoff_time = datetime.utcnow().replace(tzinfo=pytz.UTC) - timedelta(hours=24)

for uid in uids:
    result, msg_data = mail.uid('fetch', uid, '(RFC822)')
    if result != 'OK':
        continue

    raw_email = msg_data[0][1]
    msg = email.message_from_bytes(raw_email)

    date_str = msg.get('Date')
    try:
        email_date = email.utils.parsedate_to_datetime(date_str)
        if email_date.tzinfo is None:
            email_date = email_date.replace(tzinfo=pytz.UTC)
    except:
        continue

    if email_date < cutoff_time:
        continue  

    subject = decode_mime_words(msg.get('Subject'))
    sender = decode_mime_words(msg.get('From'))
    recipient = decode_mime_words(msg.get('To'))

   
    body = ""
    if msg.is_multipart():
        for part in msg.walk():
            if part.get_content_type() == "text/plain" and not part.get("Content-Disposition"):
                charset = part.get_content_charset()
                body = part.get_payload(decode=True).decode(charset or 'utf-8', errors='ignore')
                break
    else:
        charset = msg.get_content_charset()
        body = msg.get_payload(decode=True).decode(charset or 'utf-8', errors='ignore')

    emails.append({
        'uid': uid.decode(),
        'to': recipient,
        'from': sender,
        'subject': subject,
        'content': body,
        'timestamp': email_date.strftime('%Y-%m-%d %H:%M:%S %Z')
    })

mail.logout()


df = pd.DataFrame(emails)

df.head()



def clean_content(text):
    text = str(text)  
    text = text.replace("\r\n", " ")  
    text = re.sub(r'\s+', ' ', text) 
    return text.strip() 

def remove_url(text):
    pattern = re.compile(r'https?://\S+|www\.\S+')
    return pattern.sub('', text)


def remove_punc(text):
    return text.translate(str.maketrans('', '', string.punctuation))

def remove_stopwords(text):
    text = str(text) 
    new_text = []
    for word in text.split():
        if word.lower() in stopwords.words('english'):
            new_text.append(' ')
        else:
            new_text.append(word)
    x = new_text[:]
    new_text.clear()
    return ' '.join(x)



df['content'] = df['content'].str.lower()
df['content'] = df['content'].apply(clean_content)
df['content'] = df['content'].apply(remove_url)
df['content'] = df['content'].apply(remove_punc)
for i in range (0 ,df.shape[0]) :
    df['content'][i] = remove_stopwords(df['content'][i])
df['content'] = df['content'].apply(lambda x : emoji.demojize(x))



df['subject'] = df['subject'].str.lower()
df['subject'] = df['subject'].apply(clean_content)
df['subject'] = df['subject'].apply(remove_url)
df['subject'] = df['subject'].apply(remove_punc)
for i in range (0 ,df.shape[0]) :
    df['subject'][i] = remove_stopwords(df['subject'][i])

df['subject'] = df['subject'].apply(lambda x : emoji.demojize(x))




df.to_pickle("emails.pkl")
df.shape

df.head(10)
