==> 2_analyse.ipynb <==
import configparser
import json
import requests
import pandas as pd
from huggingface_hub import InferenceClient

file_path = 'emails.pkl'
df = pd.read_pickle(file_path)
df.head(10)



config = configparser.ConfigParser()
config.read("app.config")

type_labels = json.loads(config["analyse"]["type_labels"])
urgency_labels = json.loads(config["analyse"]["urgency_labels"])


API_URL_classification = "https://router.huggingface.co/hf-inference/models/facebook/bart-large-mnli"
headers = json.loads(config["analyse"]["headers"])

def type_classification(sequence ):
    payload = { "inputs": sequence,   "parameters": {"candidate_labels": type_labels }}
    response = requests.post(API_URL_classification, headers=headers, json=payload)
    result = response.json()
    return result['labels'][0], result['scores'][0]

def urgency_classification(sequence ):
    payload = { "inputs": sequence,   "parameters": {"candidate_labels": urgency_labels }}
    response = requests.post(API_URL_classification, headers=headers, json=payload)
    result = response.json()
    return result['labels'][0], result['scores'][0]


def sentiment(sequence ):
    client = InferenceClient(
    provider="hf-inference",
    api_key= config['analyse']['api_key']
)
    result = client.text_classification(
    text=sequence,
    model="cardiffnlp/twitter-roberta-base-sentiment-latest"
)
    return dict(result[0])['label'] , dict(result[0])['score']


for i in range( 0 , df.shape[0]):

    sequence = sequence = df['content'][i]
    max_length = 600
    if len(sequence) > max_length:
        sequence = sequence[:max_length]

    class_type , class_confidence = type_classification(sequence)
    urgency_label , urgency_confidence = urgency_classification(sequence)
    sentiment_label , sentiment_score = sentiment(sequence)

    df.loc[i, 'type'] = class_type
    df.loc[i, 'type_confidence'] = class_confidence
    df.loc[i, 'sentiment'] = sentiment_label
    df.loc[i, 'sentiment_confidence'] = sentiment_score
    df.loc[i, 'urgency'] = urgency_label
    df.loc[i, 'urgency_confidence'] = urgency_confidence

    print(f'row {i} done')

df.head(20)


df.to_pickle("type.pkl")

df.head(10)
