==> 4_upload.ipynb <==
import pandas as pd
from google.cloud import bigquery
from google.oauth2 import service_account
from pandas_gbq import to_gbq
from datetime import datetime
import configparser
import json

df = pd.read_pickle('emails.pkl')
df2 = pd.read_pickle('type.pkl')

project_id = 'fetchemail-3005'

now = datetime.now()
timestamp_str = now.strftime("%Y%m%d_%H%M%S")
table_id1 = f'fetchemail-3005.Logs.email_fetch_{timestamp_str}'
table_id2 = f'fetchemail-3005.Logs.email_type_{timestamp_str}'

config = configparser.ConfigParser()
config.read("app.config")
service_account_credentials_path = config['upload']['service_account_credentials_path']



try:
  
    credentials = service_account.Credentials.from_service_account_file(service_account_credentials_path)

    if not credentials.scopes:
        credentials = credentials.with_scopes(
            ["https://www.googleapis.com/auth/bigquery"]
        )

    
    client = bigquery.Client(credentials=credentials, project=project_id)


    to_gbq(df, table_id1, project_id=project_id, if_exists='replace', credentials=credentials)
    to_gbq(df2, table_id2, project_id=project_id, if_exists='replace', credentials=credentials)
    print(f"DataFrame written to BigQuery table: {table_id1} , {table_id2}using OAuth 2.0 credentials from: {service_account_credentials_path}")

except Exception as e:
    print(f"An error occurred during BigQuery upload with OAuth 2.0 credentials: {e}")
